requirements.txt
dbt-spark==1.7.0
pyspark==3.4.1
delta-spark==2.4.0
boto3==1.34.0
python-dotenv==1.2.1
